Revert "moved fragmentation and library spectrum to new project Omics"
This reverts commit d1bc75c6aaccaf08c3d2c31318e1a6e58ea6684c.
