--- mzLib/Omics/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace Omics
-{
-    public class Class1
-    {
-
-    }
-}
