+++ mzLib/Proteomics/PSM/PsmFromTsv.cs
@@ -1,5 +1,6 @@
 ﻿using Chemistry;
 using Easy.Common.Extensions;
+using Proteomics.Fragmentation;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
@@ -8,8 +9,6 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using Omics.Fragmentation;
-using Omics;
 
 namespace Proteomics.PSM
 {
