+++ mzLib/Proteomics/Protein/Protein.cs
@@ -1,9 +1,9 @@
-﻿using Proteomics.ProteolyticDigestion;
+﻿using Proteomics.Fragmentation;
+using Proteomics.ProteolyticDigestion;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using Omics.Fragmentation;
 
 namespace Proteomics
 {
