+++ mzLib/Proteomics/ProteolyticDigestion/PeptideWithSetModifications.cs
@@ -1,13 +1,13 @@
 ﻿ using Chemistry;
 using MassSpectrometry;
 using Proteomics.AminoAcidPolymer;
- using System;
+using Proteomics.Fragmentation;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
- using Omics.Fragmentation;
 
- namespace Proteomics.ProteolyticDigestion
+namespace Proteomics.ProteolyticDigestion
 {
     [Serializable]
     public class PeptideWithSetModifications : ProteolyticPeptide
