+++ mzLib/Test/DatabaseTests/TestProteomicsReadWrite.cs
@@ -4,8 +4,8 @@ using System.IO;
 using System.Linq;
 using MassSpectrometry;
 using NUnit.Framework;
-using Omics.Fragmentation;
 using Proteomics;
+using Proteomics.Fragmentation;
 using Proteomics.ProteolyticDigestion;
 using UsefulProteomicsDatabases;
 using Stopwatch = System.Diagnostics.Stopwatch;
