+++ mzLib/Test/TestLibrarySpectrum.cs
@@ -1,12 +1,11 @@
 ﻿using NUnit.Framework;
+using Proteomics.Fragmentation;
 using Proteomics.PSM;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Omics;
-using Omics.Fragmentation;
 
 namespace Test
 {
