+++ mzLib/Test/TestPeptides.cs
@@ -20,11 +20,11 @@ using Chemistry;
 using MzLibUtil;
 using NUnit.Framework;
 using Proteomics.AminoAcidPolymer;
+using Proteomics.Fragmentation;
 using Proteomics.ProteolyticDigestion;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Omics.Fragmentation;
 using UsefulProteomicsDatabases;
 using Stopwatch = System.Diagnostics.Stopwatch;
