+++ mzLib/Test/TestProductMassesMightHaveDuplicates.cs
@@ -2,11 +2,11 @@
 using MassSpectrometry;
 using NUnit.Framework;
 using Proteomics;
+using Proteomics.Fragmentation;
 using Proteomics.ProteolyticDigestion;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Omics.Fragmentation;
 using Stopwatch = System.Diagnostics.Stopwatch;
 
 namespace Test
